Fixes and remaining tests: ✓

+++ src/phg/sfm/sfm_utils.h
@@ -8,7 +8,7 @@ namespace phg {
 
     void randomSample(std::vector<int> &dst, int max_id, int sample_size, uint64_t *state);
 
-    bool epipolarTest(const cv::Vec2d &pt0, const cv::Vec2d &pt1, const cv::Matx33d &F, double t);
+    bool epipolarTest(const cv::Vec2d &pt0, const cv::Vec2d &pt1, const cv::Matx33d &F, long double t);
